Rename configureAxiosDefaults to configureGlobalAxiosDefaults and remove duplicated binaryToBuffer from core
