+++ packages/@n8n/backend-network/src/http/binary-string.ts
@@ -1,29 +1,9 @@
 import { IncomingMessage } from 'http';
 import { decode } from 'iconv-lite';
-import { UnexpectedError } from 'n8n-workflow';
 import type { Readable } from 'stream';
 
+import { binaryToBuffer } from './binary-buffer';
 import { parseIncomingMessage } from './parse-incoming-message';
-
-/** Converts a readable stream to a buffer */
-async function streamToBuffer(stream: Readable): Promise<Buffer> {
-	return await new Promise<Buffer>((resolve, reject) => {
-		const chunks: Buffer[] = [];
-		stream.on('data', (chunk: Buffer) => chunks.push(chunk));
-		stream.on('end', () => resolve(Buffer.concat(chunks)));
-		stream.once('error', (cause) => {
-			if ('code' in cause && cause.code === 'Z_DATA_ERROR')
-				reject(new UnexpectedError('Failed to decompress response', { cause }));
-			else reject(cause);
-		});
-	});
-}
-
-/** Converts a buffer or a readable stream to a buffer */
-async function binaryToBuffer(body: Buffer | Readable): Promise<Buffer> {
-	if (Buffer.isBuffer(body)) return body;
-	return await streamToBuffer(body);
-}
 
 /**
  * Decodes a buffer or stream into a string, resolving the character encoding
