+++ packages/core/src/binary-data/__tests__/utils.test.ts
@@ -1,38 +1,9 @@
 import { UnexpectedError } from 'n8n-workflow';
 import { Readable } from 'node:stream';
-import { createGunzip } from 'node:zlib';
 
-import { binaryToBuffer, createFixedSizeChunker } from '@/binary-data/utils';
+import { createFixedSizeChunker } from '@/binary-data/utils';
 
 describe('BinaryData/utils', () => {
-	describe('binaryToBuffer', () => {
-		it('should handle buffer objects', async () => {
-			const body = Buffer.from('test');
-			expect((await binaryToBuffer(body)).toString()).toEqual('test');
-		});
-
-		it('should handle valid uncompressed Readable streams', async () => {
-			const body = Readable.from(Buffer.from('test'));
-			expect((await binaryToBuffer(body)).toString()).toEqual('test');
-		});
-
-		it('should handle valid compressed Readable streams', async () => {
-			const gunzip = createGunzip();
-			const body = Readable.from(
-				Buffer.from('1f8b08000000000000032b492d2e01000c7e7fd804000000', 'hex'),
-			).pipe(gunzip);
-			expect((await binaryToBuffer(body)).toString()).toEqual('test');
-		});
-
-		it('should throw on invalid compressed Readable streams', async () => {
-			const gunzip = createGunzip();
-			const body = Readable.from(Buffer.from('0001f8b080000000000000000', 'hex')).pipe(gunzip);
-			const promise = binaryToBuffer(body);
-			await expect(promise).rejects.toThrow(UnexpectedError);
-			await expect(promise).rejects.toThrow('Failed to decompress response');
-		});
-	});
-
 	describe('createFixedSizeChunker', () => {
 		const drain = async (source: Readable): Promise<Buffer[]> => {
 			return await new Promise((resolve, reject) => {
