+++ packages/core/src/binary-data/index.ts
@@ -1,4 +1,4 @@
 export * from './binary-data.service';
 export { BinaryDataConfig } from './binary-data.config';
 export type * from './types';
-export { isStoredMode as isValidNonDefaultMode, FileLocation, binaryToBuffer } from './utils';
+export { isStoredMode as isValidNonDefaultMode, FileLocation } from './utils';
