+++ packages/core/src/binary-data/object-store.manager.ts
@@ -1,3 +1,4 @@
+import { binaryToBuffer } from '@n8n/backend-network';
 import { Service } from '@n8n/di';
 import fs from 'node:fs/promises';
 import type { Readable } from 'node:stream';
@@ -5,7 +6,6 @@ import { v4 as uuid } from 'uuid';
 
 import { ObjectStoreService } from './object-store/object-store.service.ee';
 import type { BinaryData } from './types';
-import { binaryToBuffer } from './utils';
 
 @Service()
 export class ObjectStoreManager implements BinaryData.Manager {
