+++ packages/core/src/execution-engine/node-execution-context/utils/request-helpers/index.ts
@@ -1,6 +1,6 @@
-import { configureAxiosDefaults } from '@n8n/backend-network';
+import { configureGlobalAxiosDefaults } from '@n8n/backend-network';
 
 // Applies n8n's global axios defaults and registers the request interceptor.
-configureAxiosDefaults();
+configureGlobalAxiosDefaults();
 
 export { getRequestHelperFunctions } from './factory';
